+++ Bakeries.DataAccess/Repo/IRepo/IPurchasesRepo.cs
@@ -10,6 +10,6 @@ namespace Bakeries.DataAccess.Repo.IRepo
     public interface IPurchasesRepo : IRepoBase<PurchaseModel>
     {
         Task<StockModel> GetStockDetailsFromItemId(int itemId, clsDbContext dbContext);
-        Task UpdateStockOnPurchase(int stockId, float Quantity);
+        Task UpdateStockOnPurchase(int stockId, decimal Quantity);
     }
 }
